BCDA-7113: Ensure database connections are returned when request times out (#135)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7113

## 🛠 Changes

- Run database queries with cancellable request context when possible

## ℹ️ Context for reviewers

In #133 we are switching to a singleton DB connection pool. However,
during performance testing, we noticed that connections were never
returned to the pool if requests timed out.

To avoid this, we need to pass a cancelable context into each database
query call. This PR does the following:

- For every database call (`Connection.X`), swap it out with
`Connection.WithContext(ctx)` if it is ever called during an API
request.
- For database calls that are _only_ used during unit testing or CLI
commands, don't do anything.
- For every instance where we need a context, update the function
parameters to accept a context object. This will usually be a
`r.Context()` in requests, and `context.Background()` (empty context) in
tests.

Note that in rare scenarios, we are running within an "API" context but
don't have the request context available. In these 2 cases, I created a
new context with a hardcoded timeout of 5 seconds.

## ✅ Acceptance Validation

[See the concurrency test results
here](https://jira.cms.gov/browse/BCDA-7113?focusedId=3199142&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-3199142).