BCDA-7118, BCDA-7121: Remove unused okta code (#138)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7118
https://jira.cms.gov/browse/BCDA-7121

## 🛠 Changes

- Remove /authn endpoints that are intended for Okta MFA integration
- Remove unused environment variables and methods related to Okta and
MFA tokens

## ℹ️ Context for reviewers

The Okta integration was copied over from the original BCDA application
in 2019. There was initially some intention to integrate with Okta that
was never completed, and we moved away from this and deleted the code
from the BCDA application. However, it was not deleted from the SSAS
codebase.

I have confirmed that these endpoints have not been called in the last
year, and that the integrations were never completed (since the required
environment variables do not exist.)

## ✅ Acceptance Validation

- Unit Tests are passing

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.