BCDA-7119: Add test coverage for main SSAS file (#144)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7119

## 🛠 Changes

- Adds testing for the main.go file

## ℹ️ Context for reviewers

This file is the main entrypoint for the program and all related
scripts. In order to provide better test coverage over these lines of
code, I had to restructure the methods a bit to rely less on global
package variables and move most of the main() function into separate,
smaller functions that could take in a Flags object.

I also split the start() function so that we could test the
"CreateServers" functionality without actually starting web servers. I
think we could _probably_ write a test that actually covers starting the
web servers and hitting them with HTTP requests but I had some trouble
figuring that out, so I left that alone for now. This should still cover
a good amount of the file.