BCDA-7415: Update SSAS to use ctx logger (#166)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7415

## 🛠 Changes

- TrackingIDs have been removed as args for some functions since the ID
generated was not consistent and `transaction id` (uuid) is the agreed
upon name convention and type for correlating logs.
- `Op:<func>` has been kept in for consistency but open to updating
- `JSONError` no logger generates a log statement to eliminate duplicate
logging where it is typically called
- `JSONError` status description uses const instead of free text
- a couple of tests were updated to use table driven structures
 
## ℹ️ Context for reviewers

Breaking these logging updates into smaller PRs. SSAS currently has a
substantial amount of duplicate and context-less logs. In an effort to
reduce errors being reported multiple times over and to make them more
informative, this PR focuses on:
- surfacing errors at the top level function call
- logging errors at the top level function call
- removing some duplicate errors (this PR is already fairly large, so
more removal will happen in another pr)
- reducing the usage of the `Operation<called|failed|etc>` and using the
context logger so we have a standard of logging

I will have one more follow up PR to remove the
`Operation<called|failed|etc>` altogether from the remaining instances
in the repo. I left that out to reduce an already large PR, but happy to
include if reviewers request.

**There are several things that I will be ticketing for the upcoming
sprint:**
- there are errors that occur at the ORM level, and instead of utilizing
the built-in types, we do a catch-all statement and report an error that
may be not be correct. Ex: "system not found" may be reported when the
connection failed or transaction had to roll back, etc.
- we can reduce the number of free-text errors by introducing custom
error types and also constants for holding error strings. there are use
cases for both in the SSAS app.
- as a second part of the above bullet, we can return more informative
error messages when they occur by check the error type (once we create
some error types) for 4XX type responses or more specifically, 400 level
responses. We should be returning which part of the request is invalid
so consumers are not blocked until they make a GG post.

## ✅ Acceptance Validation

updated tests, ran locally, tests pass. Will be posting screenshots of
local logs vs lower env log morning of 3/22.

## 🔒 Security Implications

- [ ] This PR adds a new software dependency or dependencies.
- [ ] This PR modifies or invalidates one or more of our security
controls.
- [ ] This PR stores or transmits data that was not stored or
transmitted before.
- [ ] This PR requires additional review of its security implications
for other reasons.

If any security implications apply, add Jason Ashbaugh (GitHub username:
StewGoin) as a reviewer and do not merge this PR without his approval.