BCDA-8240: Update dependencies to eliminate snyk vulnerabilities (#181)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8240

## 🛠 Changes

Updated bcda-ssas-app to golang 1.19.12, as well as the alpine images
and requisite dependencies to images supporting 1.22, as those are more
recently utilized.

## ℹ️ Context

Added binutils-gold in a dockerfile to ensure swagger can be built.
<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
Changes were made in order to reduce the number of snyk vulnerabilities,
including eliminating some of the critical vulnerabilities shown at:
https://app.snyk.io/org/bcda/project/b185bf9a-b24d-4ac1-a892-6759ab92bc56
and
https://app.snyk.io/org/bcda/project/2830ab17-cf13-4780-a2f2-51259a8526d0

If any of the following security implications apply, this PR must not be
merged without Stephen Walter's approval. Explain in this section and
add @SJWalter11 as a reviewer.
- Adds a new software dependency or dependencies. -> binutils-gold added
to Dockerfile.ssas

## 🧪 Validation

make test -> passes
Passes snyk tests and CI tests on github.