Fix calls to fmt Errorf with no interpolation (#187)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8373

## 🛠 Changes

Convert various fmt.Errorf statements to either require the
interpolation or to errors.New. Both versions should return a string as
an error.

## ℹ️ Context

The build was failing on golangci-lint checks.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Linting and testing were run to verify no breaking changes.