BCDA-8360 add waf sync lambda (#197)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8360

## 🛠 Changes

Add WAF auto sync lambda

## ℹ️ Context

We have recently switched over to WAF to control our allowlist of IP
addresses. This lambda should sync our app DB with the source of truth
on IP addresses to our WAF.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local tests and linting passing.

---------

Co-authored-by: Parwinder Bhagat <Parwinder.Bhagat@e14s.com>