BCDA-9877: Add comment to IP address checker (#208)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8977

## 🛠 Changes

Added a comment to explain some confusing code related to our IP address
validator.

## ℹ️ Context

Our IP address validator looks like it should accept CIDR ranges but
does not, nor do we want it to allow CIDR ranges as that could be a
potential security issue.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.