BCDA-9092: log cms ID when credentials are reset/revoked, or system is created (#218)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9092

## 🛠 Changes

Added CMS identifier in logs when:
- credentials are reset 
- credentials are revoked
- a new system is created

Updated tests for all of the above to ensure the CMS ID is present.

## ℹ️ Context

Added for better visibility around requests to rotate or create
credentials.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests updated and passing.