BCDA-9145: temporarily create two ci-checks workflows (#222)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9145

## 🛠 Changes

Create two versions of ci-checks: one for PRs and one for workflow calls
from the bcda-app repository.
This is a temporary measure until we can give access to ssas codebuild
runners from bcda-app.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->