BCDA-9359: Cleanup docker images (#242)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9359

## 🛠 Changes

Clean up of docker images. Removed unnecessary packages/statements,
updated for best practices, and added non-root user.

## ℹ️ Context

In preparation for fargate we want to clean up, update, and secure our
docker images. Overall we have the following dockerfiles:
- package: Builds an RPM that is used for our EC2 instance. Can likely
go away with move to fargate.
- ssas: Creates OpenAPI docs. Creates SSAS compiled application and API
config.
- tests: Sets up an image usable for testing and CI suite (gosec,
linter, etc).
- postman_test: Sets up a SSAS image and runs postman tests against it.

I did not see a useful way to combine or reduce this number for now.  

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local builds.