BCDA-9375: Minor cleanup for WAF sync lambda (#243)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9375

## 🛠 Changes

Two minor changes:
- Dont log IP addresses (not useful)
- Alert if updated IP addresses is ever empty (could potentially be an
issue)

## ℹ️ Context

With the recent incident with WAF Sync lambda we had some thoughts on
updating WAF sync lambda in various ways. This is just a small part of
that cleanup.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local testing, workflow integration testing.

---------

Co-authored-by: Parwinder Bhagat <Parwinder.Bhagat@e14s.com>