BCDA-9361: Add logging to stdout (#244)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9361

## 🛠 Changes

Add feature flag enabled version of logging that goes to STD OUT instead
of writing to a file.

## ℹ️ Context

Fargate migration will need the ability to write to STD OUT instead of
writing to files. Feature flag so that we can be dynamic in changing
between these two configurations.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting, testing.