BCDA-9505: remove tokens.go init (#253)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9505

## 🛠 Changes

remove init in tokens.go; initializing needed variables in config
package. Moved init to function `SetAccessTokenCreator` that is called
in main.

## ℹ️ Context

Changes were made as larger initiative to remove init in the codebase.
Init has been a development slowdown and blocker when troubleshooting
singular tests.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests passing.