Carl 9263 adjust workflows (#259)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9263

## 🛠 Changes

Corrected the publish workflow to account for an issue when running the
CI checks workflow multiple times. (Same work as was done in
`bcda-app`).

## ℹ️ Context

New fargate workflows were duplicating some unnecessary work and was
also causing failures.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Workflow run.