BCDA-9570: fix publish workflow (#263)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9570

## 🛠 Changes

- update publish workflow to checkout ssas repo instead of bcda-app repo
- remove an unnecessary field from docker compose

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

When the publish workflow is run, it should build and publish the
code/docker image contained in this repo

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
Will test after merge, due to how we use this workflow from the bcda-app
repo.