BCDA-9624: Add sleep to SSAS smoke tests for concurrent request verif… (#266)
…ication

## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9624

## 🛠 Changes

Add a small sleep to make sure concurrent but linked requests dont fail
due to race condition.

## ℹ️ Context

This smoke test was failing once we scaled SSAS up to more than one
instance.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Smoke test run after this is merged.