BCDA-9624: Move smoke test sleep to prerequest script (#268)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9624

## 🛠 Changes

Move sleep in smoke test to prerequest.

## ℹ️ Context

It seems like previously the request would happen, then it would sleep,
then it would process. Which is very much not the intended flow desired.
Now it will process the prerequest logic (sleep), then send the request.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Workflow test.