BCDA-9583: Update ssas publish to work per env (#270)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9583

related PRs:
https://github.com/CMSgov/bcda-ops/pull/1293
https://github.com/CMSgov/bcda-app/pull/1271

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->

- update publish workflow to push to an environment-specific ECR
repository

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

In https://github.com/CMSgov/bcda-ops/pull/1293 the ECR repositories are
updated to be environment-specific. This PR updates the publish workflow
to push to the new repo based on which environment is being deployed to.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

tested identical changes to sibling workflows in bcda-app by publishing
to dev