BCDA-9587: Resolve duplicate loggers (#278)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9587

## 🛠 Changes

Simplify routers to remove duplicate logger instances.

## ℹ️ Context

Previously there was a shared base router that instantiated an API
logger. Then each individual router would instantiate its own API logger
again. This was the simplest approach to resolve that issue. This should
also simplify things a bit, even if it does end up with some additional
lines of code.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.